Moved GeoJSON tests into its own file
